speech-model/requirements.txt
# Core ML frameworks
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0

# Database
psycopg2-binary>=2.9.0

# Visualization & experimentation
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
tensorboard>=2.13.0

# Evaluation
scikit-learn>=1.3.0
jiwer>=3.0.0  # Word Error Rate calculation

# API Service
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

